src/requirements.txt
python-docx>=0.8.11
docx2txt>=0.8
PyQt6>=6.4.0
